requirements.txt
aiohttp==3.8.2
aiosignal==1.3.1
alembic==1.11.1
algomojo==1.4
alpaca==1.0.0
alpaca-py==0.8.2
alpaca-trade-api==3.0.2
annotated-types==0.5.0
appdirs==1.4.4
APScheduler==3.10.1
apturl==0.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
avro==1.11.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blinker==1.4
Brlapi==0.7.0
bs4==0.0.1
certifi==2023.7.22
certipy==0.1.3
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==41.0.2
cssselect==1.2.0
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
deprecation==2.1.0
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
exceptiongroup==1.1.2
exchange-calendars==4.2.8
fake-useragent==1.1.3
fastavro==1.8.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.2
frozendict==2.3.8
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
httplib2==0.14.0
idna==2.8
importlib-metadata==6.8.0
importlib-resources==6.0.0
install==1.3.5
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.4
jsonschema-specifications==2023.6.1
jupyter-telemetry==0.1.0
jupyterhub==4.0.1
keyring==18.0.1
korean-lunar-calendar==0.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
lxml==4.9.3
macaroonbakery==1.3.1
Mako==1.2.4
MarkupSafe==2.1.3
msgpack==1.0.3
multidict==5.2.0
multitasking==0.0.11
nest-asyncio==1.5.7
netifaces==0.10.4
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.24.4
oauthlib==3.2.2
olefile==0.46
outcome==1.2.0
packaging==23.1
pamela==1.1.0
pandas==1.5.3
pandas-ta==0.3.14b0
parse==1.19.1
pexpect==4.6.0
Pillow==7.0.0
pkgutil-resolve-name==1.3.10
prometheus-client==0.17.1
protobuf==3.6.1
psycopg2-binary==2.9.6
py4j==0.10.9.7
pyarrow==12.0.1
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
pydantic==1.10.12
pydantic-core==2.4.0
pyee==8.2.2
PyGObject==3.36.0
PyJWT==1.7.1
pyluach==2.2.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==23.2.0
pyppeteer==1.0.2
pyquery==2.0.0
pyRFC3339==1.1
PySocks==1.7.1
pyspark==3.4.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
pyxdg==0.26
PyYAML==6.0
referencing==0.29.3
regex==2023.6.3
reportlab==3.5.34
requests==2.31.0
requests-file==1.5.1
requests-html==0.10.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rpds-py==0.8.11
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
SecretStorage==2.3.1
selenium==4.10.0
sgmllib3k==1.0.0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
sseclient-py==1.7.2
ssh-import-id==5.10
systemd-python==234
tinysegmenter==0.3
tldextract==3.4.4
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
tweepy==4.14.0
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.16
w3lib==2.1.1
wadllib==1.3.3
webdriver-manager==4.0.0
webencodings==0.5.1
websocket-client==1.6.1
websockets==10.4
wsproto==1.2.0
xkit==0.0.0
yarl==1.9.2
yfinance==0.2.22
zipp==3.16.2
